+++ src/index.ts
@@ -2,6 +2,7 @@ export * from "./components/Badge/Badge";
 export * from "./components/Button/Button";
 export * from "./components/Card/Card";
 export * from "./components/Checkbox/Checkbox";
+export * from "./components/Dialog/Dialog";
 export * from "./components/Input/Input";
 export * from "./components/Loading/Loading";
 export * from "./components/Tabs/Tabs";